+++ internal/command/testdata/test/with_verify_module/verify/main.tf
@@ -1,0 +1,11 @@
+variable "id" {
+  type = string
+}
+
+data "test_data_source" "resource_data" {
+  id = var.id
+}
+
+resource "test_resource" "another_resource" {
+  id = "hi"
+}